+++ changelog.md
@@ -7,6 +7,7 @@
 - A filter to allowing the overwritting of the `banner_navigation` shortcode.  `to_banner_navigation_atts`
 - Added in an option to switch the itinerary to the destination images instead of the accommodation images.
 - A filter to allow the use of destination pins in the map, for rail and boat tours. `lsx_to_get_itinerary_ids_meta_key`
+- A function to allow you to include the parent images in the itinerary pool. `lsx_to_itinerary_append_parent_destinations`
 
 ### Fixed
 - The banner easing navigation.